Added function matching + fixed #include syntax.
