+++ .gitignore
@@ -371,11 +371,6 @@ healthchecksdb
 *.a
 *.lib
 
-# Executables
-*.exe
-*.out
-*.app
-
 **/cmake-build-debug
 **/CMakeCache.txt
 **/cmake_install.cmake
