+++ src/P2/game.c
@@ -1,6 +1,6 @@
-#include <game.h>
-#include <joy.h>
-#include <gs.h>
+#include "game.h"
+#include "joy.h"
+#include "gs.h"
 
 /* Sets the number of coins on the save file */
 void SetCcoin(int nParam)
