+++ tools/match.cmd
@@ -1,0 +1,2 @@
+@echo off
+CodeMatcher.exe sly_proto.elf -compile C:\sly1\src\P2
