+++ tools/matching_guide.txt
@@ -1,0 +1,26 @@
+__Code Matching Guide for Sly1 Decompilation Project__
+Works only on Windows for now.
+
+**1. Clone the Sly1 Git repo.**
+`git clone https://github.com/TheOnlyZac/sly1.git`
+
+**2. Download and extract the PS2 compiler collection (25MB in size).**
+`https://archive.org/download/SNSystemsProDGPs2/ProDGPs2usrLocalSceFiles.zip`
+Extract it to your C:\ drive.
+Once extracted, the path structure should look like: `C:\usr\local\sce\...`
+If you were wondering why this step is needed, it will be useful for compiling our code (from the repo).
+CodeMatcher will be able to match it against Sly1's original code.
+
+**3. Get the Sly1 Prototype ELF.**
+Find the Sly1 August Prototype ISO online, download it and extract the ELF to the tools folder in the repo. Or ask on Discord.
+We cannot provide this file via the Git repo for copyright reasons.
+
+**4. Decompile, code, sleep, repeat.**
+Run match.cmd from the tools folder in the repo to compile the source tree and perform a function match.
+A summary of the process will be displayed, including functions that do match and which do not.
+
+If you have any questions, feel free to leave a message on Discord!
+
+Have fun!
+
+fenik0/felinis
